Add exception handlers in user service

+++ src/main/java/com/autentia/sharedexpenses/shared_expenses_app/Services/UserService.java
@@ -2,6 +2,7 @@ package com.autentia.sharedexpenses.shared_expenses_app.Services;
 
 import com.autentia.sharedexpenses.shared_expenses_app.Domain.User;
 import com.autentia.sharedexpenses.shared_expenses_app.Repository.IUserRepository;
+import com.autentia.sharedexpenses.shared_expenses_app.Services.Exceptions.UserNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -32,33 +33,49 @@ public class UserService {
 
     //Find a user by their id:
 
-    public Optional<User> getUserById(long id){
-        return userRepository.findById(id);
-    }
+    public Optional<User> getUserById(long id) {
 
+        if (userRepository.existsById(id)) {
+            return userRepository.findById(id);
+        } else {
+            throw new UserNotFoundException();
+        }
+    }
 
     //Update user:
 
     public User updateUser(User user, long id){
 
-        User userToUpdate = userRepository.findById(id).orElseThrow(() -> new RuntimeException("User not found"));;
+        if (userRepository.existsById(id)) {
 
-        userToUpdate.setName(user.getName());
-        userRepository.updateUser(userToUpdate, id);
-        return userToUpdate;
+            User userToUpdate = userRepository.findById(id).orElseThrow(UserNotFoundException::new);
+
+            userToUpdate.setName(user.getName());
+            userRepository.updateUser(userToUpdate, id);
+            return userToUpdate;
+        } else {
+            throw new UserNotFoundException();
+        }
     }
 
     //Delete user:
 
     public Boolean deleteUser(long id){
 
-        try {
-            userRepository.deleteById(id);
-            return true;
+        if (userRepository.existsById(id)) {
 
-        } catch (Exception e){
-            return false;
+            try {
+                userRepository.deleteById(id);
+                return true;
+
+            } catch (Exception e){
+                return false;
+            }
+
+        } else {
+            throw new UserNotFoundException();
         }
+
     }
 
 }
